[crypto][kimchi] compute circuit digest via compiled gates
We previously computed the hash of a circuit using its intermediate representation, before it was compiled down to gates (with permutation info).
This meant that different circuits could potentially lead to the same hash.
This commit makes us compute the digest based on the compiled gates, using code from proof-systems.