feat: Update design to include potential of dynamic lookups
Dynamic lookup of array structures in SnarkyJS is not implemented yet
and this feature can drastically change the design of the RFC and
implementation once merged. This commit makes a note that this feature
might be included and needs to be considered more once it lands. The
assumption of building a general purpose JSON parser is the main talking
point behind dynamic lookup.