Fix MinaSigner unit tests
Makes the 'didShutdown' variable and the 'shutdown' function global
inside the MinaSigner. The correct usage of the shutdown function should
be that it is only called once. The way it was used was wrong and was
called multiple times in each test suite, resulting in test failures.
Instead, we added a 'setupTests.ts' file that runs the shutdown function
at the end of each test suite so we only call shutdown once. This
results in all the unit tests passing properly.