Fix chain & peer reliability tests' flakiness
Problem: both chain reliability and peer reliability tests rely on
blocks to be produced within an expected time interval. However when a
node is stopped, this remains a valid assumption no longer. When a
significant portion of stake is offline, block creation may not happen
naturally and make the test fail with legitimate reasons for it.

Solution: remove stake from the node that is being stopped.