Split up kimchi bindings
This splits up kimchi bindings into 3 stages:
* `kimchi_types`, the pure OCaml types that the rust functions consume
  and return;
* `pasta_bindings`, the bindings for field and group elements, for which
  we have equivalents in snarkette;
* `kimchi_bindings`, the bindings for functions related to the kimchi
  proof system.

The intention is that this splitting will make it far easier to remove
pickles -- and thus the rust backend -- as a dependency for everything
up to the blockchain/transaction snark modules. Once this is done, it
should be straightforward to create a virtual module system for the
pasta curve and field types, and thus to compile the client SDK directly
from `mina_base` and friends, without further need for optional
compilation.