Package upload more atomic
During package uploads, packages are now uploaded as temporary filenames
and then moved to their final names in bulk.

This prevents:
1. corrupted files with visible paths if the transfer is interrupted
2. a user downloading a corrupted file while it is being transferred

Ultimately, this improves (but cannot completely fix) the possibility of
race conditions between deb-s3 and other jobs downloading from the
repository.

Manifests are uploaded as before, since they have to be uploaded as soon
as possible, before the lock is released.