kimchi_bindings: set cargo target dir to _build/cargo_kimchi_*
Previously, we were letting cargo autodetect a cargo directory,
which was dependent on the workspace. dune would not copy
src/lib/crypto/Cargo.toml, and so cargo would find the
mina root workspace (/Cargo.toml) from the source directory.

However, we shouldn't rely on dune not copying this file,
or use the target directory from this workspace.

Now, we calculate the path _build/cargo_kimchi_{wasm,bindgen,stubs},
and provide it to cargo via $CARGO_TARGET_DIR. This simplifies some
code and allows for more determinism.