Use kad-dht patcher, drop custom connection logic
Problem: custom connection logic which drops connections after 400ms
when the high water is reached leads to many problems because many
protocols in Mina's networking stack rely on an ability to establish a
short-lived connection.

This custom logic was introduced as a hacky solution to an issue in
kad-dht library which lead to 75% of peers to become connected and
connections to them never being trimmed.

Solution: remove the custom connection logic and use Patcher to
curcumvent the 75% issue.