Fix performance of downloading bodies
Problem: asymptotical behaviour of starting `Downloading_body`
state handling is O(n) because of the need to traverse `n`
states deep to find failed `Downloading_body` states.

Solution: introduce a `next_failed_ancestor` field and use it to
store the closest known ancestor that failed once.

This trick is similar to path compression of Disjoint-set data structure
and author of the commit expects `O(logn)` performance, although no
proof was constructed to justify this expectation.

Crucially, when there was no failed `Downloading_body` in ancestry,
performance will be plain `O(1)`.