Make hashing over transactions snarks partial (ledger hashes)
NOTE: `work_id` is still computed by hashing the full statement,
only the result of `hash` changes.

Hashing the full statement is computationally expensive, with
the consequence of making the transaction snark refcount tables
costly to maintain.

The assumption here is that the source+target ledger hashes contained
in the statement is still unique enough.