test-executive: Add zkApp payment to ensure they don't drop
After sending the first two transactions, the first being one that
should be dropped and then the second being one that is accepted, we
want to ensure additional zkApp payments are processed and not dropped
if the first one fails. Here, we add an additional zkApp payment
following the second to ensure it doesn't get dropped.