Use serde for ProofEvaluations to fixup WASM bindings
This commit copies the strategy of serde_wasm_bindgen, using serde to
emit JS values that match the format expected by js_of_ocaml for the
`ProofEvaluations` struct. This change may appear invasive and complex,
but the manual mapping over the internal structures in javascript proved
much more complex and error prone. This has the additional bonus that,
if any new evaluations are added, it will not be necessary to change the
WASM bindings in any way.

Note that we do not do this for the whole `ProverProof` structure. In
particular, the representation of affine group elements given by serde
is a binary one, whereas we want an `(F * F) or_infinity`. Fixing this
issue also proved very complex, so this commit walks the middle path of
only serde-serializing the `ProofEvaluations` struct, where no group
elements appear.