Don't ban nodes for block created in early catchup
Problem: when a node is in early catchup and creates a block, it might
be considered worse than root and the node to be banned for gossiping
it.

Solution: if parent of the transition is in root history, do not ban.

Additionally, if we've seen gossip with the parent of transition and
its ancestor was in root history, do not ban. This one is implemented
heuristically via an LRU cache.