Pickles: allow commitment lengths to vary
For this commit, though, we keep it constant, as it was, to keep changes
minimal.

This commit also simplifies the creation of commitment lengths by removing the
`of_int` parameter since it was always set to to `Fn.id`, that is, no conversion
from integer was ever occuring, since the committed values were always integers.