Kimchi bindings: expand num chunks computation
Follow what's been done in Kimchi and dynamically compute the number of chunks
from the index parameters. Defaults to 1 in general.

When it's not possible/easy to derive a `num_chunks` value, the current
implementation defaults to exposing a parameter in the function to be set
appropriately at the call sites.

This commit either uses a `zk_rows` variable (which for now should be set to 3
by the caller so as to ensure compatibility) or directly 3.