Kimchi bindings/Pickles: expose zk_rows in OCaml verifier index
Mimic the structure for the Kimchi-based index. `zk_rows` is clearly a parameter
of the protocol that cannot be computed elsewhere.

Define an identifier `default_zk_rows` for the fixed default value that was used
until now in module `Common`. This will make it easier to substitute it by
adequate computation in the future than having to search for the value `3`.