[transaction_logic] Fee payments as a finite state machine.
This is a proof of concept of re-implementing the transaction logic
as a finite state machine. For the moment we implement only fee
payments and plug the code into the real transaction logic to see
how it behaves in tests. If it works, we'll re-implement further
parts of the transaction logic in this manner.