In the result of a zkApp transaction application, make the ordering of the accessed accounts list explicit.
The current version implicitly depends on the ordering provided
by the `Hashtbl` implementation in the `base` library, which
is not explicit. This order my change either by changes to the
`Hashtbl` implementation, or because of changes to this function,
so it is important to make the ordering explicit and stable.