Reduce the timeout for the Nodes_to_synchronize wait condition
This wait condition usually runs very quickly, and we only see the
timeout being hit when the test infrastructure has failed. With the
previous timeout of 48 slots, this takes 1.6 hours. Reducing this
significantly should help with long-running failed tests.

The numbers here are chosen arbitrarily. Really, the time it takes
should depend on the known state of the nodes entering the wait
condition, accounting for the number of blocks to be bootstrapped /
caught-up to, but I don't want to do that in this first pass.