feat: add consensus_test.ml and dune files for consensus testing
The consensus_test.ml file is added to read block data, validate blocks, convert them into OCaml block type, initialize PoS data structures, run the PoS selection algorithm on the blocks, and print out the results. This is done to test the consensus algorithm.

The dune file is added to define the build instructions for the consensus_test.ml file. This is necessary for the OCaml build system to know how to compile and link the consensus_test.ml file.