Kimchi_backend: split the finalization
Fixes https://github.com/MinaProtocol/mina/issues/14019

The finalize_and_get_gates function is more complicated to extract from the
functor as there are dependencies with other functions like add_row.
As the finalization of the fixed and runtime lookup tables are independent from
the fields, it makes sense to expose it outside of the functor.