Fix premature stops of Orchestrator
Problem: Orchestrator gets stopped if it can't reach a node
for executing stop-daemon command or if it can't find any
participants.

Solution: don't treat no-reply on stop-daemon as an error and repeat
discovery in the loop to prevent stopping an experiment because of
glitches.

Issue: https://github.com/MinaProtocol/mina/issues/13957