Relax all ipc test requirements
The IPC tests were requiring that nodes never reconnect after the
initial connection. In the libp2p upgrade, the rules for when
connect and disconnect events are fired changed for the receiving end of
a connection. Now, a node receiving an incoming connection request will
trigger connection events before gating, and will thus trigger a
disconnect event after gating. Therefore, banning will still work, but
nodes will get a connect and a disconnect every time a banned peer
attempts to establish a connection, which fails the test unnecessarily.