Prototype of optimizing lookups over masks
Problem: 290 layers of masking trees makes processing of transactions to
be O(290*logn) where n is the size of a mask. This 290x factor is a
significant slowdown

Solution: have a mechanism to replace O(290*logn) to something around
O(30*logn).

This PR is prototype, without logic integrated into frontier handling