Remove a TODO in location_of_account_batch
Justification: we don't have a choice in this function because unlike
loading an account by location here we can't tell of an account whether
it's a new one or just wasn't loaded from parent.

It might have an implication to performance of transaction processing (I
do see some db lookups despite the fact that we preload a lot of stuff).
But I think we'll handle it via special-casing unsafe_preload
function, not right here.