refactor graphql log engine to be functor and reuse with cloud and local
The changes were done to make the graphql_polling_log_engine more reusable by making it a functor. This allows it to support different networks, such as the docker network. The old files were deleted as they are no longer needed. The error_json was added to the dune file in integration_test_lib for better error handling.