add a polling interval to graphql log functor
Changed the Make_GraphQL_Polling_log_engine to accept a polling interval
parameter. This is important because inside the start_filtered_log
function, we want to call it as soon as the graphql API is available
for a node in the docker network. If we wait longer, we have the
potential to miss the `Node_initialization` event which will lead to a
test failure. Additionally exposed the retry_delay_sec value as a
function parameter to start_filtered_log.