Allow delegation backend to run without AWS/Google
Problem: for testing networks maintaining Google Sheets and AWS storage
is a nuisance.

Solution: introduce config entries to allow running the services with no
Google sheet and AWS configured. An in-memory map of submissions that
happened within last 15 minutes is maintained and could be accessed via
`/v1/online` HTTP endpoint. Whitelist of public keys is configured in as
a fixed list of strings.