Specialize runtime_genesis_ledger
Do not allow any unnecessary variation and allow only runtime configs
that use the fields of fork config export.

Tool exits with an exception if there are some unexpected fields in the
runtime config to ensure it can't be misused. Also tar generation code
that is run within the tool was greatly simplified.