Nix: use Rocksdb 5.x for building #mainnet/#devnet
Problem: ledgers built with the latest Rocksdb aren't compatible with
the Rocksdb 5.x that is used in Docker builds. This is a problem when
nix tooling is used to prepare .tar.gz filed for ledgers (e.g. as part
of hard fork flow).

Solution: for #mainnet and #devnet flake outputs use custom rocksdb
derivation (of version 5.11.3). Using exactly the same version as is
used in Docker would be more work. because derivation for 5.11.3 was
available in nix repository's git history, while for version of Docker
it would be extra effort.

Generated ledger files are confirmed to be compatible with the Docker
build.