Upload built nix derivations to the cache
When CI provides the necessary secret env variables (or they are
provided locally), cache is updated with built binaries.

This makes sure that repretetive runs do not need to rebuild binaries.
But the most useful use case is that if `compatible` branch is updated
rarely, it will not be rebuilt between updates, reducing the runtime of
test by a few minutes at least.