Use source tree directly from nix/ocaml.nix
Reverts change introduced in fcbea58b.

In develop it isn't used. And it prevents further work in incremental
nix builds of Mina. If a patch will need to be applied to a source tree,
it's recommended to come up with a more involved process where patch
file is given as an argument to nix/ocaml.nix, then split to a number of
separate files (using patchutils) and then every patch is applied in
patch phases of derivations that work on chunks of tree.