Use mina_version wisely
Problem: Module mina_version uses mina_version.normal as the default
implementation which causes recompilation of most of the project on
every commit.

Solution:

* use mina_version.dummy as the default implementation
* rename mina_version_normal to mina_version_compiled
* use mina_version_compiled in all archive, rosetta and archive
  executables

This PR also makes mina_version_compiled a separate package (not part of
mina_version). While this is unnecessary for the purpose of PR, it eases
dependency analysis on package level vs. library level. This is useful
for some work that will be introduced in follow-up PRs.