Promote src/config/config.mlh
Problem: need to promote the config.mlh for lsp-server to work with it.

Solution: move all .mlh files to src/config/components and rewrite
src/config/dune to have one non-promoted artifact depending on
components and one promoted artifact that depends on promoted (similar
to how it was with src/config/config.mlh and src/config.mlh, but all
within src/config.