Add dedicated config generation script for compatible
At this point, the file is correctly parsed by the node so that it can start.

Still needs testing to see if we're able to reboot the network from the
generated configuration. Importing accounts from the configuration does not seem
to succed in the default time provided (around 30m).

Currently testing with expanded time constraints (giving it 2hrs).