feat: Add optional timestamp field to BlockTransaction
This commit adds an optional `timestamp` field to the `BlockTransaction` type in the `rosetta_models/block_transaction.ml` file. The `timestamp` field represents the timestamp of the block in milliseconds since the Unix Epoch. This field is not part of the official spec and it's optional, only being included if the Rosetta node is run with the `--include-timestamp` flag.