Allow to precompute transaction hashes
Context: when staged ledger diff application is performed, some hash
operations that are invoked do not depend on an underlying ledger, but
only on the transaction.

Problem: when staged ledger diff application is executed within context
of block creation, these hashing routines unnecessarily slow down the
block creation.

Solution: allow to precompute certain values for a transaction and be
supplied for staged ledger diff application.

P.S. this commit is first in series, and it doesn't use precomputed data
everywhere it could be used. Neither it updates block creation and tx
pool.