Fix broadcast_transition
Problem: broadcast_transition's purpose is to distribute a gossip
received on some topics to other topics.

Existing implementation couldn't accomodate the case of gossip being
received on both topics (in which case there is no need to broadcast it
any further).

Solution: accept a list of topics for which the item (block, tx or snark
work) was received.