Reduce repetitions in TestBitswapSmall
Existing test is really long to execute. On some machines it exceeds the limit of 60 minutes.

With the reduced number of repetitions, testing is still performed, but it's ensured that libp2p unit tests take less than 30m even on slower machines.

In future we may look deeper in the test and instead of making lesser amount of repetitions either optimize the software or restructure the test.

Tests concerns only bitswap, i.e. functionality that is not activated in mainnet's Mina build.