Polymorphize Zkapp_command in mina_wire_types
Context: zkapp command appears in two flavors in the code. One flavor
is zkapp command with intermediate hashes being computed for the call
forest. This representation is required for the zkapp command
application. Another flavor is zkapp command with call forest hashes
being replaced with dummies (of `unit` type). This is used for
transmission of a transaction over the wire.

Problem: these two flavors are defined as completely separate types
with all of the functions operating on either of them. This isn't handy
for writing code that is agnostic of flavor. Which either neccesitates
unnecessary conversions or code duplication.

Solution: make the zkapp command type definition in `mina_wire_types`
polymorphic in account updates structure. Perform similar changes in
zkapp command definition (in `mina_base` package). This paves the way
for writing polymorphic function for things that do not depend on call
forest hashes.