kimchi-bindings/js: back on nightly and use 2023-09-18
nightly has been initially removed as it seemed to make vendoring all the
dependencies more complicated because some std lib must be included.
However, removing nightly introduces issues like "DataCloneError", see
https://github.com/o1-labs/o1js/issues/1989, which makes the reliance on
nightly inevitable.
Removing --offline because stdlib missing for now (like `compiler_builtins`).

Another note is that cargo vendor has been introduced to provide a
consistent build environment for all platforms and all users, including
the CI. The lockfile Cargo.lock is meant to provide a locked environment
with all the dependencies, including the transitive ones. However, it is
only used by binaries. In our build process, we do build a static
library (libwires_15_stubs.a), not a binary. And Rust doesn't use the
lockfile in this case.
Another solution than `cargo vendor` would have been to strictly enforce
the version in the top-level Cargo.toml, but it implies to add all the
transitive dependencies (as otherwise, when a new transitive dependency
is released, cargo will fetch it, making the environment inconsistent).
This solution is not viable as the purpose of Cargo.toml is to enforce
versions on direct dependencies only, and leave the cargo solver doing
its job for the transitive dependencies.
For these reasons, the solution of vendoring the dependencies to build
the static library has been selected.