Implement tests for disk_cache.filesystem + bugfix
Implement tests via test_lib and fix a bug.

The bug was in finalizer of id object to be called after the cache
directory is removed by the unit test engine, finalizer callback
attempted to remove a file in a directory that didn't exist and an
exception being propagated all the way up the callstack.

Fix was in making the exception ignored. There is no much of handling
that we can do in the finalizer, so it's wise to simply ignore the
exception.