Remove unnecessary use of virtual modules
Proof_cache_tag was originally meant to host the on-disk cache, hence it
was implemented via the virtual modules technique. Later we decided to
move this functionality to a separate `disk_cache` module, using
`proof_cache_tag` as a specialization for proofs. This works really
well, but after implementing this change, it became obvious that we
don't need both `disk_cache` and `proof_cache_tag` to use virtual
modules.

This PR simplifies the code.