Attempt to create genesis proof at network start
The block producer would formerly attempt to create a genesis proof
immediately before it was needed in the block production iteration. This
creation has been moved outside of this iteration. The block producer
will now spawn a thread that attempts to create it immediately after it
detects that the frontier has been initialized.