Add a block verifier test subcommand to mina
The command

mina advanced test verify-block-commands

can be used to verify the commands in a precomputed block and gather
internal logs for the timing of different stages of verification. The
--block-file option indicates the json file containing the precomputed
block - by default it points to the file that is downloaded and cached
when building the mina_block tests. The --log-dir indicates the
directory to which the logs should be output - by default it is the
current working directory.