Remove Interruptible usage from block production
The former block generation loop was structured internally with
Interruptible in a way that allowed async block generation tasks to
cancel previously-spawned async block generation tasks. After the
refactoring of the block creation scheduling, this functionality became
unused, so it and the use of Interruptible itself have now been removed.