Add With_hash.Set module
A new module With_hash.Set provides a set that uses `(_, _) With_hash.t`
as elements, using hash for operating the set (and ignoring the value).

A new module is added to Mina_stdlib.
The module provides an interface for general-purpose set (which may
not necessary be implemented via Comparable's Make).