Allow for full actions, events in ledger test
The --transfer-parties-get-actions-events flag is now recognized in the
mina ledger test. When set, the command generator will fill the actions
and events of every account update of every Zkapp_command used in the
ledger test. Otherwise, these fields will be left empty. By default,
this flag is off (matching the current behaviour).