Add a Signature_transfers_spec command generator
The Signature_transfers_spec and signature_transfers command generator
in the Transaction_snark.For_tests generalizes the existing
Multiple_transfers_spec by allowing for the receivers in the spec to
have full keypairs, and by allowing for the transfer parties (sender and
receivers) to receive the actions and events from the spec in their
updates (via the transfer_parties_get_actions_events flag).

The existing Multiple_transfers_spec and multiple_transfers are now
implemented using this new command generator.