Add flag to transaction snark test spec
The transfer_parties_get_actions_events flag in the transaction snark
test spec controls whether only the zkapp accounts receive the actions
and events detailed in the spec in their updates (if false) or if the
transfer parties receive them as well (if true). This flag will be used
in a test that will be added in a later commit - it is currently always
false, matching the previous behaviour of the command generator.