Fix daemons not creating a lockfile after cleanup
If a daemon finds that a daemon lockfile exists but is not held by any
process (because the PID in the file does not correspond to any running
process) then it must delete that stale lockfile. With these changes, it
will then re-attempt to acquire the lockfile. This is important, as this
lockfile is used to exclude other daemons from starting up in the same
config directory. Before, the daemon would simply continue startup
without acquiring the lockfile.