Add an aux field to Account_update
This field (in Account_update.Poly.t and Account_update.Fee_payer.t) is
intended to contain auxiliary data related to the Account_update. For
now, the only type this field is used with is the No_aux.t type, which
indicates that no auxiliary data is present. In future, the non-stable
types will cache the hash of the actions in their body in this field.

Since the aux field is not intended for serialization, the yojson/sexp
instances for Account_update.Poly.t have been modified to preserve the
existing serialization structure. In particular, during serialization
this field is ignored, and during deserialization this field cannot be
present, and can only take on the type No_aux.t.