Add dummy Aux_data to unstable Account_update
This Aux_data will hold the cached value of the hash of the actions in
the account update body. For now, it only takes on the value Field.zero.

Various functions have been adapted to improve their polymorphism, so
they can work with both the stable and unstable versions of the account
update types.