Make sync ledger pipe synchronous
Problem: buffered pipe for bootstrap had no semantical value. It's built
on top of an already-existing buffered pipe in the router.

Solution: use synchronous pipe.The only behavior change is that some
transitions that would otherwise be dropped when boostrap would've
finished will be transfered instead to the succeeding transition
frontier pipe.