Represent choosable pipe as a pair
Represent choosable synchronous pipe as a pair of reader and writer.

Internally implementation still uses the same type `'a t` to represent
both reader and writer, but on the interface this detail is hidden.

With clear separation of reader and writer, it's made clear on type
level that closure should be attempted on writer, while read/iteration
on reader.

Implementation still initializes one ivar and returns it both as writer
and reader (and over the process of read/write these two will morph into
different values).

As a side-change commit also updates some of the tests in a minor way.