Ensure PID lock files are synced to disk
The Async.Writer.save function uses with_file_atomic to ensure that a
given string is always written to a file completely or not at all.
However, if the ?fsync parameter for that function is set to false - the
default, and what it was before these changes - the temp file that
with_file_atomic creates may not be flushed to disk before being moved
to the target location. Thus, a well-timed interruption might cause the
PID file to be created but be entirely empty.

Now, the buffer will be flushed before the file is moved, which should
ensure that the PID file is always in the correct state: filled with a
PID, or absent entirely.