app/logproc: rename public executable logproc into mina-logproc
Until now, the executable `logproc` is not unused by anyone. In the CI, the
binary is renamed into `mina-logproc`.
With this patch, the command `make install` or `dune install` will add in the
user path the binary `mina-logproc`, so the user can share tests commands with
the CI.
This is another step toward a unification of the CI and the local UX