Abstract the snarked root ledger
A new Root mina ledger type has been created, to encapsulate the
functionality of the snarked root ledger. A Root ledger is an unmasked,
fully-committed and database-backed ledger that can be used as the base
of a ledger mask tree. It is only possible to create these with a single
backing database, for now; in a future update the root definition will
be expanded so that they can be backed by a pair of stable and unstable
account databases that are kept in sync by a converting merkle tree.

This new root ledger is also used to represent the genesis ledger.