Define and use a Config.t type for Root ledgers
The Root.Config module defines how a Root ledger can be created. There
are two independently varying options:

1. The directory_name to use for the database(s) backing the Root
2. The backing_type for the Root

Some methods have been defined on the abstract Config.t type, to
represent common operations on the locations of the databases backing
root ledgers. The intent is to support existing patterns in the ledger
handling code (deleting the backing of a root ledger, testing if a
backing already exists, creating a root ledger based on a particular
directory name) while abstracting the filesystem details of the root
ledger backing. This is necessary for the future converting merkle tree
backing for root ledgers; the migrated database will need to be created,
moved, and deleted properly.