Add a flag to pad app_state
With the introduction of hardfork feature (in develop branch) of zkapp
app_state containing 32 elements, it's necessary for the
runtime_genesis_ledger to support padding the app state to necessary
length.

In future this flag will be replaced by a flag performing all the
necessary ledger transformations required for planned hardfork.