Remove transfer_accounts_with from Root
This method was supposed to be implementable in a more efficient way
with a converting database backing. However, the most convenient way of
doing that involves a different API, where the method is responsible for
creating the destination database itself. If avoiding the re-hashing
that Ledger_transfer does is found to be desirable then this new API can
be implemented and used when populating a Root from genesis or a
recovered potential snarked ledger.