Compute precise hard fork ledger information
The faster hard fork config generation method will involve applying the
diffs from a ledger mask branch onto a copy of its root. For that reason
it will be convenient to keep around more information about those
ledgers, so they can be migrated properly. The next epoch ledger is most
relevant for the upcoming hardfork; it could be

1. its own root, which should be checkpointed directly, or

2. an uncommitted ledger in the transition frontier, which should have
its accumulated diff harvested and applied to the root snarked ledger