Split genesis ledger tar creation method in two
The previous generate_tar method did two things:

1. It prepared the data needed to generate a new ledger database tar
   file, given some Ledger.t mask over a ledger database
2. It generated the actual tar files and names given that data

When we generate the tar files for the migrated ledger databases during
the hard fork, we will not have Ledger.t masks available to us. Rather,
we will have known databases of hard fork accounts. In this situation,
it is easier to do (1) separately, and then re-use (2) to generate the
actual files.

The generate_tar method now does (2) alone; a separate
generate_ledger_tar method now does (1) and (2).