Make converting ledger sync check more defensive
If a daemon is stopped at certain points while syncing a converting
ledger to a network, the underlying databases can be left in an
unexpected state: the num_accounts in the database will be set to a high
value (because the daemon added a batch of accounts at high addresses)
but not all accounts at lower indices will be present (because the
daemon didn't get to filling them in yet). This will cause functions
like iteri and get_at_index_exn to throw unexpected exceptions when they
get to addresses that are missing accounts.

A couple of new ledger functions have been added to account for this
possibility, which are used in an updated ledger database sync check
implementation. A few related unit tests have been added to the
converting merkle tree, one of which does not pass without these changes
being in place.