Increase rosetta test postgres memory threshold
The postgres instance in the rosetta tests is exceeding the current
default memory limit of 2500MB somewhat regularly; it seems as though it
was close to this limit in past tests, and is now finally exceeding it.
The limit is now temporarily 3000MB until we either decide that the
memory usage is still fine, or we find out why the memory usage is
increasing and fix that issue.