Add notion of chain state location to the daemon
The Chain_state_locations module tracks the locations on disk of the
components that are considered part of the "chain state". These chain
state component locations are subdirectories of the chain state
directory.

The chain state directory itself is currently the mina config directory,
so the chain state components are all still located at their compatible
locations at the top level of the mina config directory. Moving the
logic of computing these locations into one module will make it easier
to change how the chain state directory is computed in future hard
forks.