Add an option to control stop time randomness
The --stop-time-interval option sets the upper bound on the number of
random hours added to the --stop-time of the daemon. It is optional,
with default 9 to match the current daemon behaviour.

Setting the option to 0 will disable the randomness entirely. This is
useful for ensuring that daemons will stop at deterministic intervals.