Refactor Root to prepare for vesting adjustment
The new Root ledger implementation explicitly depends on the exposed
ledgers from Mina_ledger.Ledger. This simplifies a lot of the root
definitions, and makes declaring the root interface much easier. It is
also a functor, allowing the caller to pass in its own account
conversion function.

It is also an existential type that wraps a specific concrete ledger
implementation, so that the root ledger creation functions can
dynamically select a specific implementation when root ledgers are
created/opened. For now, the only concrete implementation is the
existing one, which assumes that the account conversion method is known
at compile time. In future, this implementation will need to change.