Rename hardfork-mode to hardfork-handling
The (now-closed) hard fork automation MIP specifies that the name of the
option controlling the feature will be --hardfork-handling, with values
keep-running and migrate-exit.

This is an internal flag; for the upcoming Mesa hard fork, users are
expected to choose the auto hard fork package if they want to use the
"auto" mode. For that reason, the specification in the MIP gave it this
somewhat awkward new name to more accurately reflect what daemon
behaviours will change if it is enabled.