Set auto fork config slot and timestamp properly
The hard fork global slot since genesis and the hard fork timestamp are
now set according to the stop slots and the hard fork genesis slot
delta, if those components of the daemon config have been set. If they
have not, then the slot and timestamp of the hard fork block itself are
used as a fallback.