Set default HF test slot time to 30s
Now that the block producer node is also running a snark worker in the
hard fork test, there seems to be enough resource contention that the
block producer node has a hard time meeting the current default 15s slot
time reliably. This can cause the hard fork test to fail, because the
block producer node's best tip can end up far enough ahead of the snark
worker nodes' best tip that the resulting desync causes fork validation
to fail.

The 30s time appears to be generous enough for the test to succeed
reliably. This is the new default for both the pre-fork chain and the
post-fork chain. These parameters can still be controlled via the
--main-slot and --fork-slot options in the hard fork test on an ad-hoc
basis as necessary.