Introduce multi-key file storage
It's a very simple way to store a set of related values in an efficient
manner.

Values are assumed to written all in one batch. Values can't be updated
or deleted individually (the only way possible is to overwrite the whole
file or remove the file).

Reading of the key reads only the bytes of file that are reuired for
the read.

It's an efficient way to store data that is semantically connected,
written and garbage-collected, but that needs to be accessed piecewise.